Cleaned up conversion warning
git-svn-id: svn://localhost:7744/trunk/alps@4253 3665984a-f8b0-4bb5-b30c-cdcaecbe7ced

+++ tool/maxent_simulation.cpp
@@ -90,7 +90,7 @@ void MaxEntSimulation::dostep()
   spex_str << "\n";
   for (int a=0; a<chi_sq.size(); ++a) 
     chisq_str << alpha[a] << " " << chi_sq[a] << std::endl;
-  double a_chi = 0;
+  int a_chi = 0;
   double diff = std::abs(chi_sq[0]-ndat());
   for (int a=1; a<chi_sq.size(); ++a) {
     double diff_new = std::abs(chi_sq[a]-ndat());
